requirements.txt
datasets>=3.3.2,<4.0.0
ipykernel>=6.29.5,<7.0.0
pandas>=2.2.3,<3.0.0
python-dotenv>=1.0.1,<2.0.0
google-genai>=1.3.0,<2.0.0
checkov>=3.2.379,<4.0.0
langchain-text-splitters>=0.3.6,<0.4.0
langchain-community>=0.3.19,<0.4.0
langgraph>=0.3.9,<0.4.0
langchain-google-genai>=2.1.0,<3.0.0
langchain[mistralai]>=0.3.20,<0.4.0
langchain-mistralai>=0.2.7,<0.3.0
langchain-core>=0.3.45,<0.4.0
python-frontmatter>=1.1.0,<2.0.0
chromadb>=0.6.3,<0.7.0
tqdm>=4.67.1,<5.0.0
langchain-chroma>=0.2.2,<0.3.0
langchain-huggingface>=0.1.2,<0.2.0
sentence-transformers==2.4.0